telegram-review-bot/pyproject.toml
[tool.poetry]
name = "telegram-review-bot"
version = "0.1.0"
description = "A Telegram bot to parse review tags in code archives."
authors = ["isofinly <isofinly.business@icloud.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.8,<3.14"
pyTelegramBotAPI = "^4.14.0"
patool = "^1.12.0"
py7zr = "^0.22.0"
python-dotenv = "^1.0.0"
rarfile = "^4.0"
tqdm = "^4.66.1"
minio = "^7.2.0"

[tool.poetry.scripts]
telegram_review_bot = "src.bot:run_bot"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
